+++ src/commons/generatorCellNumber.js
@@ -1,0 +1,7 @@
+const generateColumnLetter = colIndex => (
+  String.fromCharCode(97 + colIndex).toUpperCase()
+);
+
+export default (index, rowNumber) => (
+  `${generateColumnLetter(index)}${rowNumber}`
+);
