+++ src/formatters/cells/generatorNumberCell.js
@@ -1,0 +1,3 @@
+import generatorCellNumber from '../../commons/generatorCellNumber';
+
+export default (index, value, rowIndex) => (`<c r="${generatorCellNumber(index, rowIndex)}"><v>${value}</v></c>`);
