+++ src/formatters/cells/generatorStringCell.js
@@ -1,0 +1,4 @@
+import escape from 'lodash.escape';
+import generatorCellNumber from '../../commons/generatorCellNumber';
+
+export default (index, value, rowIndex) => (`<c r="${generatorCellNumber(index, rowIndex)}" t="inlineStr"><is><t>${escape(value)}</t></is></c>`);
