+++ src/formatters/rows/generatorRows.js
@@ -1,0 +1,7 @@
+import formatRow from './formatRow';
+
+export default rows => (
+  rows
+  .map((row, index) => formatRow(row, index))
+  .join('')
+);
