+++ src/statics/[Content_Types].xml.js
